// src/types/api-response.type.ts
export interface APICharacterDTO {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface APIResponseDTO {
  info?: {
    count: number;
    pages: number;
    next?: string;
    prev?: string;
  };
  results: APICharacterDTO[];
}